fix(limit-order): dedupe cancelled order notifications
Move MyOrders shared row and empty-state pieces into feature components/utils. Track cancelled notifications only after the initial snapshot so replayed data does not emit duplicate analytics.
