+++ apps/kyberswap-interface/src/components/LimitOrder/MyOrders/index.tsx
@@ -1,18 +1,17 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { Trans, t } from '@lingui/macro'
-import { HTMLAttributes, ReactNode, useCallback, useEffect, useMemo, useState } from 'react'
-import { Trash } from 'react-feather'
+import { t } from '@lingui/macro'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useSearchParams } from 'react-router-dom'
 import { useGetListOrdersQuery } from 'services/limitOrder'
 
-import { ReactComponent as NoDataIcon } from 'assets/svg/no_data.svg'
-import { ButtonOutlined } from 'components/Button'
-import DropdownMenu, { MenuOption } from 'components/DropdownMenu'
+import DropdownMenu, { type MenuOption } from 'components/DropdownMenu'
 import CancelOrderModal from 'components/LimitOrder/CancelOrder/CancelOrderModal'
 import { useCancellingOrders } from 'components/LimitOrder/CancelOrder/useCancellingOrders'
 import { useLimitOrderContext } from 'components/LimitOrder/LimitOrderContext'
 import OrderRow from 'components/LimitOrder/MyOrders/OrderRow'
 import TableHeader from 'components/LimitOrder/MyOrders/TableHeader'
+import { CancelAllButton, EmptyOrders, TabSelector } from 'components/LimitOrder/MyOrders/components'
+import { useMyOrdersNotifications } from 'components/LimitOrder/MyOrders/useMyOrdersNotifications'
 import {
   LIST_ORDER_TABS,
   PAGE_SIZE,
@@ -24,7 +23,6 @@ import {
 import { useLimitOrderTracking } from 'components/LimitOrder/hooks/useLimitOrderTracking'
 import { LimitOrder, LimitOrderStatus } from 'components/LimitOrder/types'
 import { isActiveStatus } from 'components/LimitOrder/utils'
-import Loader from 'components/Loader'
 import Pagination from 'components/Pagination'
 import RefetchIndicator from 'components/RefetchIndicator'
 import SearchInput from 'components/SearchInput'
@@ -35,66 +33,10 @@ import { useInvalidateTagLimitOrder } from 'hooks/useInvalidateTags'
 import usePageLocation from 'hooks/usePageLocation'
 import useTab from 'hooks/useTab'
 import { sortChainOptionsByPriority } from 'pages/Earns/hooks/useSupportedDexesAndChains'
-import { cn } from 'utils/cn'
-import {
-  subscribeNotificationOrderCancelled,
-  subscribeNotificationOrderExpired,
-  subscribeNotificationOrderFilled,
-} from 'utils/firebase'
 import { isSupportLimitOrder } from 'utils/index'
-
-type NotificationOrderCallback = Parameters<typeof subscribeNotificationOrderExpired>[2]
 
 const ALL_CHAINS_VALUE = 'all'
 const EMPTY_LIMIT_ORDERS: LimitOrder[] = []
-
-const NoResultWrapper = ({ className, ...rest }: HTMLAttributes<HTMLDivElement>) => (
-  <div
-    className={cn(
-      'flex min-h-[220px] flex-col items-center justify-center gap-2 text-sm font-medium text-subText',
-      className,
-    )}
-    {...rest}
-  />
-)
-
-type TabSelectorProps = {
-  activeTab: LimitOrderStatus
-  rightContent?: ReactNode
-  setActiveTab: (n: LimitOrderStatus) => void
-}
-
-const TabSelector = ({ activeTab, rightContent, setActiveTab }: TabSelectorProps) => (
-  <div className="flex min-w-0 items-center border-b border-darkBorder">
-    <div className="flex min-w-0 flex-1 items-center overflow-x-auto" role="tablist">
-      {LIST_ORDER_TABS.map((tab, index) => {
-        const active = tab === activeTab
-        const isLast = index === LIST_ORDER_TABS.length - 1
-        return (
-          <button
-            key={tab}
-            aria-selected={active}
-            className={cn(
-              'relative flex min-h-11 shrink-0 cursor-pointer items-center gap-1.5 border-0 px-4 py-3 text-sm font-medium',
-              !isLast && 'border-r border-darkBorder',
-              active
-                ? 'bg-transparent text-primary hover:bg-transparent hover:text-primary'
-                : 'bg-transparent text-subText hover:bg-transparent hover:text-text',
-            )}
-            onClick={() => setActiveTab(tab)}
-            role="tab"
-            type="button"
-          >
-            <span className="text-base font-medium leading-[normal]" style={{ color: 'inherit' }}>
-              {tab === LimitOrderStatus.ACTIVE ? <Trans>Active Orders</Trans> : <Trans>Order History</Trans>}
-            </span>
-          </button>
-        )
-      })}
-    </div>
-    {rightContent && <div className="flex shrink-0 items-center px-4">{rightContent}</div>}
-  </div>
-)
 
 const MyOrders = () => {
   const { account } = useActiveWeb3React()
@@ -230,6 +172,8 @@ const MyOrders = () => {
     } catch (error) {}
   }, [onReset, refetchOrders])
 
+  useMyOrdersNotifications({ account, chainId, limitOrderTracking, networkName, refreshListOrder })
+
   const setKeyword = useCallback(
     (val: string) => {
       setSearchParams(getSearchParamsWithKeyword(searchParams, val), { replace: true })
@@ -292,47 +236,6 @@ const MyOrders = () => {
     setIsCancelAll(true)
   }
 
-  const trackCancelledOrder = useCallback(
-    (order: LimitOrder) => {
-      limitOrderTracking.trackMyOrderCancelled(order, networkName)
-    },
-    [limitOrderTracking, networkName],
-  )
-
-  const trackFilledOrder = useCallback(
-    (order: LimitOrder) => {
-      limitOrderTracking.trackMyOrderFilled(order, networkName)
-    },
-    [limitOrderTracking, networkName],
-  )
-
-  useEffect(() => {
-    if (!account) return
-
-    const callback: NotificationOrderCallback = data => {
-      const orders: LimitOrder[] = data?.orders ?? []
-      if (orders.length) refreshListOrder()
-    }
-
-    const unsubscribeCancelled = subscribeNotificationOrderCancelled(account, chainId, data => {
-      refreshListOrder()
-      const cancelledOrders: LimitOrder[] = data?.orders ?? []
-      cancelledOrders.forEach(trackCancelledOrder)
-    })
-    const unsubscribeExpired = subscribeNotificationOrderExpired(account, chainId, callback)
-    const unsubscribeFilled = subscribeNotificationOrderFilled(account, chainId, data => {
-      const filledOrders: LimitOrder[] = data?.orders ?? []
-      if (filledOrders.length) refreshListOrder()
-      filledOrders.forEach(trackFilledOrder)
-    })
-
-    return () => {
-      unsubscribeCancelled?.()
-      unsubscribeExpired?.()
-      unsubscribeFilled?.()
-    }
-  }, [account, chainId, refreshListOrder, trackCancelledOrder, trackFilledOrder])
-
   useEffect(() => {
     onReset()
   }, [orderType, onReset])
@@ -348,20 +251,20 @@ const MyOrders = () => {
     onReset()
   }, [orderTab, onReset])
 
-  const cancelAllButton = showCancelAll && (
-    <ButtonOutlined
-      onClick={onCancelAllOrder}
-      disabled={disabledCancelAll}
-      className={cn('w-fit gap-1.5 px-3 py-1 text-sm', !disabledCancelAll && '!border-red-35 !text-red')}
-    >
-      {isLoadingCancelAllOrders ? <Loader size="14px" /> : <Trash size={14} />}
-      <Trans>Cancel All</Trans>
-    </ButtonOutlined>
-  )
-
   return (
     <div className="flex w-full flex-col">
-      <TabSelector setActiveTab={onSelectTab} activeTab={activeTab} rightContent={cancelAllButton} />
+      <div className="flex min-w-0 items-center border-b border-darkBorder">
+        <TabSelector setActiveTab={onSelectTab} activeTab={activeTab} />
+        {showCancelAll && (
+          <div className="flex shrink-0 items-center px-4">
+            <CancelAllButton
+              onClick={onCancelAllOrder}
+              disabled={disabledCancelAll}
+              isLoading={isLoadingCancelAllOrders}
+            />
+          </div>
+        )}
+      </div>
 
       <div className="flex justify-between gap-2 px-4 py-2 max-sm:flex-col">
         <div className="flex min-w-0 items-center gap-2 max-sm:w-full">
@@ -389,6 +292,7 @@ const MyOrders = () => {
           onChange={onChangeKeyword}
         />
       </div>
+
       <TableHeader isActiveTab={isTabActive} />
       <div className="relative h-0">
         <RefetchIndicator visible={isFetching} />
@@ -416,20 +320,7 @@ const MyOrders = () => {
           />
         </div>
       )}
-      {showNoOrders && (
-        <NoResultWrapper>
-          <NoDataIcon />
-          <span>
-            {keyword ? (
-              <Trans>No orders found.</Trans>
-            ) : isTabActive ? (
-              <Trans>You don&apos;t have any open orders yet.</Trans>
-            ) : (
-              <Trans>You don&apos;t have any order history.</Trans>
-            )}
-          </span>
-        </NoResultWrapper>
-      )}
+      {showNoOrders && <EmptyOrders isActiveTab={isTabActive} keyword={keyword} />}
 
       <CancelOrderModal
         isOpen={isOpenCancel}
